Extract time parsing logic Move time parsing logic to a separate module to separate concerns
